refactor(oxfmt): fix up code and comments (#23770)
There's still room for refactoring, but this will serve as a starting point for now.
